p3_commit

Struct ExtensionMmcs

pub struct ExtensionMmcs<F, EF, InnerMmcs> { /* private fields */ }

A wrapper to lift an MMCS from a base field F to an extension field EF.

ExtensionMmcs allows committing to and opening matrices over an extension field by internally using an MMCS defined on the base field. It works by flattening each extension field element into its base field coordinates for commitment, and then reconstructing them on opening.

Implementations

impl<F, EF, InnerMmcs> ExtensionMmcs<F, EF, InnerMmcs>

pub const fn new(inner: InnerMmcs) -> Self