Crate p3_commit

A framework for various (not necessarily hiding) cryptographic commitment schemes.

Modules

testing

Structs

BatchOpening

A Batched opening proof.

BatchOpeningRef

A reference to a batched opening proof.

ExtensionMmcs

A wrapper to lift an MMCS from a base field F to an extension field EF.

LagrangeSelectors

Given a PolynomialSpace, S, and a subset R, a Lagrange selector P_R is a polynomial which is not equal to 0 for every element in R but is equal to 0 for every element of S not in R.

Traits

Mmcs

A “Mixed Matrix Commitment Scheme” (MMCS) is a generalization of a vector commitment scheme.

Pcs

A polynomial commitment scheme, for committing to batches of polynomials defined by their evaluations over some domain.

PolynomialSpace

Fixing a field, F, PolynomialSpace<Val = F> denotes an indexed subset of F^n with some additional algebraic structure.

Type Aliases

OpenedValues

OpenedValuesForMatrix

OpenedValuesForPoint

OpenedValuesForRound

Val